backend/requirements.txt
# Django Framework
Django==5.1.4
django-ninja==1.3.0
django-ninja-extra==0.21.6
django-cors-headers==4.6.0
gunicorn==23.0.0
uvicorn[standard]==0.34.0

# Database
psycopg2-binary==2.9.10

# LangGraph & LangChain
langgraph==0.2.60
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14

# Vanna AI
vanna[chromadb,openai,postgres]==0.7.5

# Web Search
tavily-python==0.5.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
pandas==2.2.3

# Async support
asgiref==3.8.1
